test: isolate hook-dependent simple fees and hip1195 suites
